refactor : DI를 사용한 비지니스 로직과 트랙잭션 코드 분리
