+++ src/main/java/service/UserServiceImpl.java
@@ -1,0 +1,37 @@
+package service;
+
+import java.util.List;
+
+import dao.UserDao;
+import pojo.User;
+
+public class UserServiceImpl implements UserService {
+    private UserDao userDao;
+    private UserLevelUpgradePolicy userLevelUpgradePolicy;
+
+    public void setUserDao(UserDao userDao) {
+        this.userDao = userDao;
+    }
+
+    public void setUserLevelUpgradePolicy(UserLevelUpgradePolicy userLevelUpgradePolicy) {
+        this.userLevelUpgradePolicy = userLevelUpgradePolicy;
+    }
+
+    @Override
+    public void upgradeLevels() {
+        final List<User> users = userDao.getAll();
+        for (User user : users) {
+            if (userLevelUpgradePolicy.can(user)) {
+                userLevelUpgradePolicy.upgrade(user);
+            }
+        }
+    }
+
+    @Override
+    public void add(User user) {
+        if (user.getLevel() == null) {
+            user.setLevel(Level.BASIC);
+        }
+        userDao.add(user);
+    }
+}
