+++ src/main/java/service/UserServiceTx.java
@@ -1,0 +1,37 @@
+package service;
+
+import org.springframework.transaction.PlatformTransactionManager;
+import org.springframework.transaction.TransactionStatus;
+import org.springframework.transaction.support.DefaultTransactionDefinition;
+
+import pojo.User;
+
+public class UserServiceTx implements UserService {
+    private UserService userService;
+    private PlatformTransactionManager transactionManager;
+
+    public void setUserService(UserService userService) {
+        this.userService = userService;
+    }
+
+    public void setTransactionManager(PlatformTransactionManager transactionManager) {
+        this.transactionManager = transactionManager;
+    }
+
+    @Override
+    public void upgradeLevels() {
+        final TransactionStatus status = transactionManager.getTransaction(new DefaultTransactionDefinition());
+        try {
+            userService.upgradeLevels();
+            transactionManager.commit(status);
+        } catch (RuntimeException ex) {
+            transactionManager.rollback(status);
+            throw ex;
+        }
+    }
+
+    @Override
+    public void add(User user) {
+        userService.add(user);
+    }
+}
