+++ src/test/java/service/UserServiceTest.java
@@ -11,6 +11,7 @@ import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
+import org.springframework.transaction.PlatformTransactionManager;
 
 import dao.UserDao;
 import pojo.User;
@@ -26,7 +27,7 @@ public class UserServiceTest {
     @Autowired
     MockMailSender mailSender;
     @Autowired
-    UserLevelUpgradePolicy levelUpgradePolicy;
+    PlatformTransactionManager transactionManager;
 
     List<User> users; // 테스트 픽스처
 
@@ -91,20 +92,22 @@ public class UserServiceTest {
         final TestUserLevelUpgradePolicy policy = new TestUserLevelUpgradePolicy(users.get(3).getId());
         policy.setUserDao(userDao);
         policy.setMailSender(mailSender);
-        userService.setUserLevelUpgradePolicy(policy);
+
+        final UserServiceImpl testUserServiceImpl = new UserServiceImpl();
+        testUserServiceImpl.setUserDao(userDao);
+        testUserServiceImpl.setUserLevelUpgradePolicy(policy);
+
+        final UserServiceTx txUserService = new UserServiceTx();
+        txUserService.setUserService(testUserServiceImpl);
+        txUserService.setTransactionManager(transactionManager);
 
         users.forEach(user -> userDao.add(user));
 
         try {
-            userService.upgradeLevels();
+            txUserService.upgradeLevels();
             Assert.fail("TestUserServiceException expected");
         } catch (TestUserServiceException ex) {
 
-        } finally {
-            final GeneralUserLevelUpgradePolicy originPolicy = new GeneralUserLevelUpgradePolicy();
-            originPolicy.setUserDao(userDao);
-            originPolicy.setMailSender(mailSender);
-            userService.setUserLevelUpgradePolicy(originPolicy);
         }
 
         checkLevelUpgraded(users.get(1), false);
